move fileByCid to courseActions and fix useSaveDraft
