+++ src/_reactComponents/Course/CourseActions.jsx
@@ -57,6 +57,48 @@ export const fileByCid = atomFamily({
   
 })
 
+export const fileByPageId = atomFamily({
+  key:"fileByPageId",
+  default: selectorFamily({
+    key:"fileByPageId/Default",
+    get:(doenetId)=> async ()=>{
+      if (!doenetId){
+        return "";
+      }
+      // const local = localStorage.getItem(doenetId);
+      // if (local){ return local}
+      try {
+        const server = await axios.get(`/media/byPageId/${doenetId}.doenet`); 
+        return server.data;
+      } catch (error) {
+        //TODO: Handle 404
+        // Error 😨
+        if (error.response) {
+          /*
+          * The request was made and the server responded with a
+          * status code that falls out of the range of 2xx
+          */
+          console.log(error.response.data);
+          console.log(error.response.status);
+          console.log(error.response.headers);
+        } else if (error.request) {
+            /*
+            * The request was made but no response was received, `error.request`
+            * is an instance of XMLHttpRequest in the browser and an instance
+            * of http.ClientRequest in Node.js
+            */
+            console.log(error.request);
+        } else {
+            // Something happened in setting up the request and triggered an Error
+            console.log('Error', error.message);
+        }
+              return "Error Loading";
+        }
+    }
+  })
+  
+})
+
 const peopleAtomByCourseId = atomFamily({
   key:"peopleAtomByCourseId",
   default:[],
